+++ bookkeeping/WingspanBookkeeping/Utils/Utilities.cs
@@ -40,6 +40,7 @@ namespace WingspanBookkeeping.Utils
             {
                 new IsoDateTimeSerializer(),
                 new EnumSerializer(),
+                new FlexibleObjectDeserializer()
             };
         }